+++ x/auth/testutil/expected_keepers_mocks.go
@@ -9,7 +9,6 @@ import (
 	reflect "reflect"
 
 	types "github.com/cosmos/cosmos-sdk/types"
-	types0 "github.com/cosmos/cosmos-sdk/x/consensus/types"
 	gomock "github.com/golang/mock/gomock"
 	protoiface "google.golang.org/protobuf/runtime/protoiface"
 )
@@ -134,42 +133,4 @@ func (m *MockAccountsModKeeper) SendModuleMessageUntyped(ctx context.Context, se
 func (mr *MockAccountsModKeeperMockRecorder) SendModuleMessageUntyped(ctx, sender, msg interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendModuleMessageUntyped", reflect.TypeOf((*MockAccountsModKeeper)(nil).SendModuleMessageUntyped), ctx, sender, msg)
-}
-
-// MockConsensusKeeper is a mock of ConsensusKeeper interface.
-type MockConsensusKeeper struct {
-	ctrl     *gomock.Controller
-	recorder *MockConsensusKeeperMockRecorder
-}
-
-// MockConsensusKeeperMockRecorder is the mock recorder for MockConsensusKeeper.
-type MockConsensusKeeperMockRecorder struct {
-	mock *MockConsensusKeeper
-}
-
-// NewMockConsensusKeeper creates a new mock instance.
-func NewMockConsensusKeeper(ctrl *gomock.Controller) *MockConsensusKeeper {
-	mock := &MockConsensusKeeper{ctrl: ctrl}
-	mock.recorder = &MockConsensusKeeperMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockConsensusKeeper) EXPECT() *MockConsensusKeeperMockRecorder {
-	return m.recorder
-}
-
-// Params mocks base method.
-func (m *MockConsensusKeeper) Params(arg0 context.Context, arg1 *types0.QueryParamsRequest) (*types0.QueryParamsResponse, error) {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Params", arg0, arg1)
-	ret0, _ := ret[0].(*types0.QueryParamsResponse)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// Params indicates an expected call of Params.
-func (mr *MockConsensusKeeperMockRecorder) Params(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Params", reflect.TypeOf((*MockConsensusKeeper)(nil).Params), arg0, arg1)
 }
